Turntable rotation and snapping working.
Managed to get a close approximation of the Ultraleap turntable working in OpenXR with XRI Toolkit.  Also managed to use the socket interactor working with mounting a part on the turntable.  Next up is to code the visuals for the turntable based on how the ultraleap team did it in their SDK.