This might get wiped off.
So tried to take the Ultraleap code and make it work here, but even though the visuals work, they use the tracking in the hand and detect if it falls within the current area, this is over an above what is really needed.  I'm going to branch and rewrite to instantiate a prefab with collider attached a fixed amount of times to see if it pans out.  There could be optimisation problems with so many colliders being used.