Hover effect completed.
Hover effect working, had to swap all materials for a blue instead of going for the glow that I wanted due to how the glow effect works.  Audio seems to trigger when the model is exploded when it really shouldn't.  Need to figure that one out.  Managed to successfully keep the far interactor disabled when a model is attached to the turntable, should make selecting individual components easier.